+++ qa/evil-tests/src/test/java/org/elasticsearch/plugins/InstallPluginCommandTests.java
@@ -51,6 +51,7 @@ import org.elasticsearch.common.cli.UserError;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.env.Environment;
 import org.elasticsearch.test.ESTestCase;
+import org.elasticsearch.test.PosixPermissionsResetter;
 import org.junit.BeforeClass;
 
 @LuceneTestCase.SuppressFileSystems("*")
@@ -61,24 +62,6 @@ public class InstallPluginCommandTests extends ESTestCase {
     @BeforeClass
     public static void checkPosix() throws IOException {
         isPosix = Files.getFileAttributeView(createTempFile(), PosixFileAttributeView.class) != null;
-    }
-
-    /** Stores the posix attributes for a path and resets them on close. */
-    static class PosixPermissionsResetter implements AutoCloseable {
-        private final PosixFileAttributeView attributeView;
-        final Set<PosixFilePermission> permissions;
-        public PosixPermissionsResetter(Path path) throws IOException {
-            attributeView = Files.getFileAttributeView(path, PosixFileAttributeView.class);
-            assertNotNull(attributeView);
-            permissions = attributeView.readAttributes().permissions();
-        }
-        @Override
-        public void close() throws IOException {
-            attributeView.setPermissions(permissions);
-        }
-        public void setPermissions(Set<PosixFilePermission> newPermissions) throws IOException {
-            attributeView.setPermissions(newPermissions);
-        }
     }
 
     /** Creates a test environment with bin, config and plugins directories. */
@@ -103,7 +86,7 @@ public class InstallPluginCommandTests extends ESTestCase {
             }
         }
     }
-    
+
     static String writeZip(Path structure, String prefix) throws IOException {
         Path zip = createTempDir().resolve(structure.getFileName() + ".zip");
         try (ZipOutputStream stream = new ZipOutputStream(Files.newOutputStream(zip))) {
@@ -382,7 +365,7 @@ public class InstallPluginCommandTests extends ESTestCase {
         Files.createFile(binDir.resolve("somescript"));
         String pluginZip = createPlugin("fake", pluginDir);
         try (PosixPermissionsResetter binAttrs = new PosixPermissionsResetter(env.binFile())) {
-            Set<PosixFilePermission> perms = new HashSet<>(binAttrs.permissions);
+            Set<PosixFilePermission> perms = binAttrs.getCopyPermissions();
             // make sure at least one execute perm is missing, so we know we forced it during installation
             perms.remove(PosixFilePermission.GROUP_EXECUTE);
             binAttrs.setPermissions(perms);
